Tests for useChangeEffect
